WIP Gameplay
This is likely going to be a content vomit.

I've discovered that I leave much sooner than I thought for this trip
and will be stretching to get all the work in as soon as possible.

If it looks complex, it's likely because I am sleep deprived letting the
haze take over.